Fix timing issue in linedrawer
This issue was not apparent when drawing only in one spot, but showd
itself when we try to clear the buffer. We set writeEnable=1 in init2
state, before the x and y registers were fully initialized. This meant
that it wrote to the framebuffer one cycle too early, writing the
previous values in the x and y registers. And since writeEnable was not
high in init3, it was set in the draw state, one cycle too late again,
so it does not draw the first pixel either. This should fix this by
setting writeEnable high fist in the init3 state.